refactor : 외부 호출 api가 트랜젝션에 묶여있지 않도록 변경
